MamboLite/requirements.txt
pandas
requests
pyyaml
pywin32; platform_system == "Windows"
